zenject試作

+++ RoomHack-Ver3/Assets/Yori/Script/System/Save/Data/PlayerSaveData.cs
@@ -4,7 +4,7 @@ public class PlayerSaveData
     public int score_Stage;
     public int score_DestoryEnemy;
     public int gunNo;
-    public float maxHitPoint;
+    public int maxHitPoint;
     public float moveSpeed;
     public float maxRamCapacity;
     public float RamRecovery;
